Add async, and fix structure. Backward compatibility is 95% preserved
