+++ contiguity_base/__init__.py
@@ -1,0 +1,12 @@
+from contiguity_base.api import AsyncContiguity, Contiguity
+from contiguity_base.base import BaseContiguity
+
+Util = BaseContiguity.Util
+
+__all__ = [
+    "AsyncContiguity", 
+    "Contiguity", 
+    "BaseContiguity", 
+    "Util"
+]
+__version__ = "1.2.0"
