+++ contiguity_base/api/__init__.py
@@ -1,0 +1,4 @@
+from contiguity_base.api.async_contiguity import AsyncContiguity
+from contiguity_base.api.contiguity import Contiguity
+
+__all__ = ["AsyncContiguity", "Contiguity"]
