+++ contiguity_base/base/__init__.py
@@ -1,0 +1,3 @@
+from contiguity_base.base.contigiuity import BaseContiguity
+
+__all__ = ["BaseContiguity"]
