+++ pyproject.toml
@@ -1,62 +1,47 @@
-[build-system]
-requires = [
-    "setuptools>=61",
-]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "contiguity_base"
-dynamic = ["version"]
-description = "The official Python SDK for Contiguity Base"
+[tool.poetry]
+name = "contiguity-base"
+version = "1.2.0"
+description = ""
+authors = ["Дмитрий Бондарев <skinxedovich@mail.ru>"]
 readme = "README.md"
-license = {file = "LICENSE.txt"}
-requires-python = ">=3.9"
-authors = [
-    {name = "Contiguity", email = "help@contiguity.support"},
-]
-keywords = [
-    "python",
-    "contiguity",
-    "deta",
-    "base",
-    "database",
-]
-classifiers = [
-    "Development Status :: 3 - Alpha",
-    "Intended Audience :: Developers",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3",
-    "Topic :: Utilities",
-    "Typing :: Typed",
-]
-dependencies = [
-    "requests>=2.25.1,<3.0.0",
-]
 
-[project.urls]
-Repository = "https://github.com/contiguity/base-python"
+[tool.poetry.dependencies]
+python = "^3.12"
+httpx = "^0.27.2"
 
-[tool.setuptools]
-package-data = {contiguity_base = ["py.typed"]}
+[tool.poetry.group.test.dependencies]
+pytest = "^8.3.3"
+pytest-asyncio = "^0.24.0"
 
-[tool.setuptools.dynamic]
-version = {attr = "contiguity_base.__version__"}
-
-[tool.uv]
-dev-dependencies = [
-    "pre-commit~=3.8.0",
-]
+[tool.poetry.group.dev.dependencies]
+ruff = "^0.6.8"
 
 [tool.ruff]
-src = ["src"]
-line-length = 119
-target-version = "py39"
+line-length = 120
+exclude = [
+    "migrations",
+    "manage.py",
+]
 
 [tool.ruff.lint]
-select = ["ALL"]
-ignore = ["A", "D"]
+unfixable = ["B"]
+select = [
+    "F",
+    "B",
+    "E",
+    "I",
+    "UP",
+    "N",
+    "ANN",
+    "Q"
+]
+ignore = ["ANN002", "ANN003", "ANN101", "ANN102", "ANN204"]
 
-[tool.pyright]
-venvPath = "."
-venv = ".venv"
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["E402"]
+"**/{tests,docs,tools}/*" = ["E402"]
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
