--- src/contiguity_base/__init__.py
@@ -1,9 +1,0 @@
-from .base import Base, Contiguity, Util, connect
-
-__all__ = (
-    "Base",
-    "Contiguity",
-    "Util",
-    "connect",
-)
-__version__ = "1.1.0"
